Reorganize cypress tests
